checkbox_ng:commands:cli: Add Curses functions
To manage suite selection and the welcome screen.

curses is part of python standard library:
http://docs.python.org/3.2/library/curses.html

It does less than urwid but does not add a new dependency just to run
command line scripts.

However it will offer a decent way to select the whitelists the tester can
choose.

For better results, paragraphs in the welcome text should be written on a
single line, textwrap taking care of lines lenghts. Only start a new line to
effectively render a new paragraph.