checkbox-ng:commands:cli: be more resilient
This patch changes how ServiceCommand (which implements the checkbox
dbus service) is enabled. Previously the code would blindly attempt to
use SeviceCommand class and silently ignore NameError, this could lead
to issues in the implementation of ServiceCommand to go unnoticed.

The new code replaces that with an if-else check. If we managed to
import the service module (which implements the ServiceCommand) then we
blindly and unconditionally instantiate it. This way post-import runtime
errors such as NameError will not stay undetected.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>