checkbox-ng: move SelectableJobTreeNode to misc.py
I plan on changing lots of cli.py so I wanted to set aside the code that
should stay intact. In addition, the class will now be shared and will
be imported from two new tiny modules that are both imported from cli.py
so it needs to be separated.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>