checkbox-ng:launcher: add LauncherDefinition
This patch adds the LauncherDefinition. A small Config-derived class
that loads (preliminary) launcher definition files. Launchers are small
executables that use checkbox-launcher as the interpreter and within
their body, specify settings for a test run. Typical settings include
the welcome message, a mask on whitelists to offer and information on
what to do with results.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>