plainbox:commands:selftest: split into cmd+inv
This patch splits the 'selftest.py' module into two new modules,
cmd_selftest and inv_selftest. During plainbox startup only the
cmd_selftest module needs to be imported and it has minimal requirements
so very little overall code needs to be loaded.

When the 'plainbox self-test' command is executed then the inv_selftest
module is imported, bringing in all of the additional modules it
requires.

The implementation of the command code was changed slightly, resulting
in one more base class and less useless initialization code. There are
no user-visible or API changes though.

In addition, all call sites that used to import the old module are
updated to use one of the new modules instead.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>