checkbox-ng:dbus: cache Introspection data
This patch adds a simple mechanism to cache calls to .Introspect() which
returns a D-Bus introspection XML string. Since the GUI apparently calls
this a *lot* it made sense to just compute it once per class (as it is
never going to change anyway).

This seems to help a little with systems having a lot of objects.

Tested-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>