checkbox-ng: load manifest provider
This patch makes checkbox-ng use extra APIs to load the manifest
provider. This is not the best way to solve it as plainbox should really
just load all of the special providers as well but for now this will
have to do. There will be separate patches to fix the provider API next.

Fixes: https://bugs.launchpad.net/plainbox/+bug/1443895

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>