urwid_ui: get rid of SA references in the UI code
Previously UI screens were given SessionAssistant instance to reference when
they wanted to grab any detailed information about units. This means that
any major change to the SessionAssistant would require updates to urwid_ui.py.
Also, if we would like to have a different way of connecting to plainbox
(not SA) then this urwid code would not work. Above all else UI code talking
directly to _model_ is a bad idea.

This patch replaces sharing SA with urwid in favor of building a list of
(for a lack of a better word) unit proxies that contain all the info UI should
need.

TODOs still left to do are:
* remove the atrocious globals in urwid_ui.py !!!11oneone
* deduplicate generation of category name maps
  (better -> do it once, and reference, best -> supply from checkbox)

Signed-off-by: Maciej Kisielewski <maciej.kisielewski@canonical.com>