remote: Enhance run_jobs with resumed_session_info data
When the slave is about to resume a session (e.g after a sut reboot),
run_jobs is called with a non empty resumed_session_info that contains
the id of the last run job.
The only ui change is the display of the resumed job id and its outcome.

When silent ui type is used a 20s delay allows the tester to manually interrupt
the auto-resume.