config: always load configs from SNAP_DATA
This patch rearranges the code of load_configs to make sure that configs from
SNAP_DATA are properly loaded even when checkbox is invoked without launcher
(e.g. when run as the daemon). This function us called every time a session is
created so the values are reloaded even if the Daemon was not restarted.

Launcher can still override the values.