resource: Add visit_Attribute() to ResourceNodeVisitor to record manifest units
To get the list of the participating manifest units, the ResourceNodeVisitor
visit_Name method is not enough as parsing 'manifest.has_foo' only returns
'manifest'.

Adding a visit_Attribute() allows parsing of the above and returns:

an ast.Name() for 'manifest' and 'has_foo' as the node attribute (attr).
If and only if the node id is 'manifest', the resource program records the list
of seen attributes.

This list of seen raw attributes is later used by _analyze_manifest() to create
a namespaced list of manifest units.